requirements_dev.txt
requests
pyyaml
numpy
pandas
scikit-learn
scipy
matplotlib
pandarallel
simplemma
stopwordsiso

